% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{weather}
\alias{weather}
\title{renvoie une partie des donnees de weather}
\usage{
weather(df)
}
\value{
renvoie les données issues de la colonne year et de la ligne 33
}
\description{
renvoie une partie des donnees de weather
}
\examples{
weather()
}
